# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs MlCity Project",
        "description": "API MlCity",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "operationId": "68ef7a386c92d5883cb7b1cf72feddc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/users/{id}/social-account": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get social account user",
                "operationId": "a39182ea029b41eb0f8fc41a420ded60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/favorite-events": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get favorites events user",
                "operationId": "9427ef2fc2617ce17437086435c32c27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/liked-events": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get liked events user",
                "operationId": "be37b0a7dc5d40e62f2f2d2556a9bc68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/favorite-sights": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get favorites sights user",
                "operationId": "3bc33b6c3b7089053c8ea8dd57a59a93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/liked-sights": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get liked sights user",
                "operationId": "275c92d082bc412bde7d1b95af4a59a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/like-event-toggle": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add event in like",
                "operationId": "8d144ed7e22b57ccc863ef9522a2ab2e",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/like-sight-toggle": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add sight in like",
                "operationId": "0ccd073ad81c75c6bf137b7c80f478ee",
                "parameters": [
                    {
                        "name": "sight_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/favorite-event-toggle": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add event in favorite",
                "operationId": "bc0b963611cdd37286a06c083824ba19",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/favorite-sight-toggle": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add sight in favorite",
                "operationId": "38d4df7a571e322a8d646ee0587d7e48",
                "parameters": [
                    {
                        "name": "sight_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/deleteUsers/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete soft user by id, method for AdminPanel",
                "operationId": "2d9ae0ccd4935df16c4d3fcb8aa2f1cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}